bugfix: rename selectors.py to not conflict with subprocess
For details see this issue comment:
https://github.com/bernardorufino/pick/issues/17#issuecomment-1029488009
